Working but not optimized

+++ src/graph/adj_lists.rs
@@ -10,17 +10,17 @@ use super::Graph;
 
 #[derive(Debug, Clone)]
 pub struct AdjLists {
-    lists: Vec<Vec<u32>>,
+    lists: Vec<Vec<usize>>,
 }
 
 impl AdjLists {
-    pub fn new(verts: u32) -> Self {
+    pub fn new(verts: usize) -> Self {
         AdjLists {
-            lists: vec![Vec::new(); verts as usize],
+            lists: vec![Vec::new(); verts],
         }
     }
 
-    pub fn from_parts<I>(verts: u32, iter: I) -> Self where
+    pub fn from_parts<I>(verts: usize, iter: I) -> Self where
         I: IntoIterator<Item = Part>
     {
         let mut graph = Self::new(verts);
@@ -30,7 +30,7 @@ impl AdjLists {
             debug_assert!(part.target_verts.end <= verts);
 
             // cannot use `part.index_to_vert()` because of borrowing rules
-            let start_index = part.owned_verts.start as usize;
+            let start_index = part.owned_verts.start;
 
             for (v, list) in part.lists.into_iter().enumerate().map(|(i, list)| (i + start_index, list)) {
                 debug_assert!(graph.lists[v].is_empty());
@@ -43,28 +43,28 @@ impl AdjLists {
 }
 
 impl Graph for AdjLists {
-    fn vertices(&self) -> Box<Iterator<Item=u32>> {
-        Box::new((0 as u32)..(self.lists.len() as u32))
+    fn vertices(&self) -> Box<Iterator<Item=usize>> {
+        Box::new((0)..(self.lists.len()))
     }
 
-    fn neighbours<'a>(&'a self, vert: u32) -> Box<Iterator<Item=u32> + 'a> {
-        Box::new(self.lists[vert as usize].iter().cloned())
+    fn neighbours<'a>(&'a self, vert: usize) -> Box<Iterator<Item=usize> + 'a> {
+        Box::new(self.lists[vert].iter().cloned())
     }
 }
 
 #[derive(Debug, Clone)]
 pub struct Part {
-    owned_verts: Range<u32>,
-    target_verts: Range<u32>,
-    lists: Vec<Vec<u32>>
+    owned_verts: Range<usize>,
+    target_verts: Range<usize>,
+    lists: Vec<Vec<usize>>
 }
 
 impl Part {
-    pub fn new(owned: Range<u32>, target: Range<u32>) -> Self {
+    pub fn new(owned: Range<usize>, target: Range<usize>) -> Self {
         assert!(owned.start <= owned.end);
         assert!(target.start <= target.end);
 
-        let len = (owned.end - owned.start) as usize;
+        let len = owned.end - owned.start;
 
         Part {
             owned_verts: owned,
@@ -73,7 +73,7 @@ impl Part {
         }
     }
 
-    pub fn add_edge(&mut self, from: u32, to: u32) {
+    pub fn add_edge(&mut self, from: usize, to: usize) {
         assert!(self.owned_verts.contains(from));
         assert!(self.target_verts.contains(to));
 
@@ -90,12 +90,12 @@ impl Part {
         let n_owned = self.owned_verts.end - self.owned_verts.start;
         let n_target = self.target_verts.end - self.target_verts.start;
 
-        let mut edges = HashSet::with_capacity(n_edges as usize);
+        let mut edges = HashSet::with_capacity(n_edges);
 
-        while edges.len() != n_edges as usize {
-            let from = rng.gen::<u32>() % n_owned;
-            let from = self.index_to_vert(from as usize);
-            let to = rng.gen::<u32>() % n_target;
+        while edges.len() != n_edges {
+            let from = rng.gen::<usize>() % n_owned;
+            let from = self.index_to_vert(from);
+            let to = rng.gen::<usize>() % n_target;
 
             edges.insert((from, to));
         }
@@ -110,12 +110,12 @@ impl Part {
     }
 
     #[inline]
-    fn vert_to_index(&self, n: u32) -> usize {
-        (n - self.owned_verts.start) as usize
+    fn vert_to_index(&self, n: usize) -> usize {
+        (n - self.owned_verts.start)
     }
 
     #[inline]
-    fn index_to_vert(&self, n: usize) -> u32 {
-        n as u32 + self.owned_verts.start
+    fn index_to_vert(&self, n: usize) -> usize {
+        n + self.owned_verts.start
     }
 }
